Create Inventory_Management.py
Dictionaries can be used to store key:value pairs.

You have been asked to create an inventory management program for a store. You use a dictionary to track all of the store's inventory along with how many of each item the store has.
store = {"Orange": 2, "Watermelon": 0, "Apple": 8, "Banana": 42} 
Complete the provided code to output the number of apples present in the store.